refactor: merge leftover economy test-support expectations (#4410)
Collapse Slice A pin/assert modules into paired scenario siblings (or a pins file when the 220-line cap would break) and move the Development panel runner into the consumer test so the support package no longer owns that assertion body.
