+++ packages/colonizethis_economy_test_support/lib/colonizethis_economy_test_support.dart
@@ -10,7 +10,6 @@ library colonizethis_economy_test_support;
 export 'src/build_cost_scenarios.dart';
 export 'src/commodity_totals_scenarios.dart';
 export 'src/boycott_blocked_commodities_scenarios.dart';
-export 'src/boycott_blocked_commodities_expectations.dart';
 export 'src/boycott_blocked_commodities_test_support.dart';
 export 'src/consumption_phases_scenarios.dart';
 export 'src/consumption_scenarios.dart';
@@ -19,17 +18,15 @@ export 'src/core_economy_test_support.dart';
 export 'src/economy_extraction_scenarios.dart';
 export 'src/economy_riches_to_treasury_scenarios.dart';
 export 'src/economy_production_scenarios.dart';
-export 'src/development_panel_read_model_expectations.dart';
 export 'src/development_panel_read_model_scenarios.dart';
 export 'src/deal_matcher_scenarios/deal_matcher_scenarios.dart';
 export 'src/frr_scenarios/frr_scenarios.dart';
-export 'src/game_lookup_helpers_expectations.dart';
 export 'src/game_lookup_helpers_scenarios.dart';
 export 'src/lock_recovery_minor_bids_scenarios.dart';
-export 'src/lock_recovery_minor_bids_expectations.dart';
 export 'src/lock_recovery_minor_bids_test_support.dart';
 export 'src/non_gp_auto_offers_scenarios.dart';
 export 'src/non_gp_auto_offers_test_support.dart';
+export 'src/non_gp_extraction_pins.dart';
 export 'src/non_gp_extraction_scenarios.dart';
 export 'src/extraction_fixture_support.dart';
 export 'src/resource_extractor_expectations.dart';
