--- packages/colonizethis_economy_test_support/lib/src/development_panel_read_model_expectations.dart
@@ -1,167 +1,0 @@
-// dart format off
-// Development panel read-model assertions (Refs #4175, #4014).
-
-import 'package:colonizethis_data/colonizethis_data.dart';
-import 'package:colonizethis_economy/colonizethis_economy.dart';
-import 'package:colonizethis_models/colonizethis_models.dart';
-import 'package:colonizethis_test/game_test_fixtures.dart';
-import 'package:colonizethis_test/test.dart';
-import 'package:colonizethis_world/colonizethis_world.dart';
-
-import 'development_panel_read_model_scenarios.dart';
-import 'extraction_fixture_support.dart';
-
-const _kWorkTargetBuildImprovement = 'build_improvement';
-
-DevelopmentPanelModel _ownedGrainImprovableModel() {
-  final map = tileMapFromGrids(
-    grid: const [
-      ['p1', 'p1'],
-    ],
-    resourceGrid: const [
-      [Resource.grain, Resource.grain],
-    ],
-  );
-  final keys = <String>[
-    for (var y = 0; y < 1; y++)
-      for (var x = 0; x < 2; x++) 'oldWorld|p1|$x|$y',
-  ];
-  final game = spainExtractorGame(
-    tileState: const TileMapState(),
-    oldWorld: RegionData(provinces: [owP1Province()]),
-    tileKeysByRegionAndProvince: {'oldWorld': {'p1': keys}},
-  );
-  return buildDevelopmentPanelModel(
-    game: game,
-    playerId: 'pl1',
-    tileMapByRegion: {'oldWorld': map},
-    topology: MapTopology(),
-    currentOrders: const Orders(),
-    provinceDisplayNamesById: const {'oldWorld|p1': 'Avalon'},
-    playerDisplayNamesById: const {'pl1': 'France'},
-  );
-}
-
-DevelopmentPanelModel _modelWithPendingWorkOrders() {
-  final game = resourceExtractorGame(tileState: const TileMapState());
-  const orders = Orders(
-    workOrdersByPlayerId: {
-      'pl1': [
-        WorkOrder(
-          unitId: 'b1',
-          target: _kWorkTargetBuildImprovement,
-          targetTileKey: 'oldWorld|p1|0|0',
-        ),
-      ],
-    },
-  );
-  return buildDevelopmentPanelModel(
-    game: game,
-    playerId: 'pl1',
-    tileMapByRegion: const {},
-    topology: MapTopology(),
-    currentOrders: orders,
-    provinceDisplayNamesById: const {},
-    playerDisplayNamesById: const {},
-  );
-}
-
-void runDevelopmentPanelReadModelExpectation(
-  DevelopmentPanelReadModelTarget target,
-) {
-  switch (target) {
-    case DevelopmentPanelReadModelTarget.ownedProvinceListsImprovableGrain:
-      final ow = _ownedGrainImprovableModel().oldWorld;
-      expect(ow.ownedScopes, hasLength(1));
-      expect(ow.ownedScopes.first.displayName, 'Avalon');
-      expect(ow.ownedScopes.first.improvableCommodities, hasLength(1));
-      expect(
-        ow.ownedScopes.first.improvableCommodities.first.commodityId,
-        CommodityCatalog.grain.id,
-      );
-      expect(ow.ownedScopes.first.improvableCommodities.first.count, 2);
-    case DevelopmentPanelReadModelTarget.modelBuildsWithPendingWorkOrders:
-      expect(
-        _modelWithPendingWorkOrders().oldWorld.idleBuilderCount,
-        greaterThanOrEqualTo(0),
-      );
-    case DevelopmentPanelReadModelTarget.playerViewFiltersUnrevealedImprovableTiles:
-      final map = tileMapFromGrids(
-        grid: const [
-          ['p1', 'p1'],
-        ],
-        resourceGrid: const [
-          [Resource.grain, Resource.grain],
-        ],
-      );
-      final keys = <String>[
-        'oldWorld|p1|0|0',
-        'oldWorld|p1|1|0',
-      ];
-      final game = spainExtractorGame(
-        tileState: const TileMapState(),
-        oldWorld: RegionData(provinces: [owP1Province()]),
-        tileKeysByRegionAndProvince: {'oldWorld': {'p1': keys}},
-      );
-      final baseView = buildPlayerView(game, MapTopology(), 'pl1');
-      final playerView = PlayerView(
-        playerId: baseView.playerId,
-        player: baseView.player,
-        ownUnitsById: baseView.ownUnitsById,
-        provincesById: baseView.provincesById,
-        visibilityByTile: const {
-          'oldWorld|p1|0|0': VisibilityLevel.fullyVisible,
-          'oldWorld|p1|1|0': VisibilityLevel.unknown,
-        },
-        prospectedTiles: baseView.prospectedTiles,
-        diplomacyByOtherId: baseView.diplomacyByOtherId,
-      );
-      final model = buildDevelopmentPanelModel(
-        game: game,
-        playerId: 'pl1',
-        tileMapByRegion: {'oldWorld': map},
-        topology: MapTopology(),
-        currentOrders: const Orders(),
-        provinceDisplayNamesById: const {'oldWorld|p1': 'Avalon'},
-        playerDisplayNamesById: const {'pl1': 'France'},
-        playerView: playerView,
-      );
-      final grain = model.oldWorld.ownedScopes.first.improvableCommodities
-          .firstWhere((r) => r.commodityId == CommodityCatalog.grain.id);
-      expect(grain.count, 1);
-      expect(grain.tileKeys, ['oldWorld|p1|0|0']);
-    case DevelopmentPanelReadModelTarget.assignedCiviliansIncludePendingBuilder:
-      final unit = Unit(
-        id: 'b1',
-        type: kUnitTypeBuilder,
-        ownerId: 'pl1',
-        locationProvinceId: 'oldWorld|p1',
-        tileKey: 'oldWorld|p1|0|0',
-      );
-      final game = TestFixtures.oldWorldGameWithUnit(unit: unit);
-      const orders = Orders(
-        workOrdersByPlayerId: {
-          'pl1': [
-            WorkOrder(
-              unitId: 'b1',
-              target: _kWorkTargetBuildImprovement,
-              targetTileKey: 'oldWorld|p1|0|0',
-            ),
-          ],
-        },
-      );
-      final model = buildDevelopmentPanelModel(
-        game: game,
-        playerId: 'pl1',
-        tileMapByRegion: const {},
-        topology: MapTopology(),
-        currentOrders: orders,
-        provinceDisplayNamesById: const {},
-        playerDisplayNamesById: const {},
-      );
-      expect(model.oldWorld.assignedCivilians, hasLength(1));
-      expect(model.oldWorld.assignedCivilians.first.unitId, 'b1');
-      expect(model.oldWorld.assignedCivilians.first.isPending, isTrue);
-  }
-}
-// dart format on
