+++ packages/colonizethis_economy_test_support/lib/src/treasury_scenarios/price_discovery_scenarios.dart
@@ -1,8 +1,38 @@
 // dart format off
 // Table-driven PriceDiscovery scenarios (Refs #3856).
 import 'package:colonizethis_economy/colonizethis_economy.dart';
+import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
-import 'price_discovery_expectations.dart';
+/// Data-driven expectations for [PriceDiscovery.computeMarketActivity] rows.
+class PriceDiscoveryMarketActivityExpectation {
+  const PriceDiscoveryMarketActivityExpectation({this.totalBidQuantity, this.totalOfferQuantity, this.filledQuantity, this.priceChangePercent, this.priceChangePercentCloseTo, this.equalsEmpty = false});
+  final int? totalBidQuantity;
+  final int? totalOfferQuantity;
+  final int? filledQuantity;
+  final double? priceChangePercent;
+  final double? priceChangePercentCloseTo;
+  final bool equalsEmpty;
+}
+void assertPriceDiscoveryMarketActivityExpectation(MarketActivity activity, PriceDiscoveryMarketActivityExpectation expectation) {
+  if (expectation.totalBidQuantity != null) {
+    expect(activity.totalBidQuantity, expectation.totalBidQuantity);
+  }
+  if (expectation.totalOfferQuantity != null) {
+    expect(activity.totalOfferQuantity, expectation.totalOfferQuantity);
+  }
+  if (expectation.filledQuantity != null) {
+    expect(activity.filledQuantity, expectation.filledQuantity);
+  }
+  if (expectation.priceChangePercent != null) {
+    expect(activity.priceChangePercent, expectation.priceChangePercent);
+  }
+  if (expectation.priceChangePercentCloseTo != null) {
+    expect(activity.priceChangePercent, closeTo(expectation.priceChangePercentCloseTo!, 1e-9));
+  }
+  if (expectation.equalsEmpty) {
+    expect(activity, equals(MarketActivity.empty));
+  }
+}
 /// One row in [priceDiscoveryNextPriceScenarios].
 typedef PriceDiscoveryNextPriceScenario = ({String label, double oldPrice, int basePrice, int newBidQuantity, int newOfferQuantity, double expectedPrice, double tolerance, String? refs});
 /// Canonical scenarios for [PriceDiscovery.computeNextPrice].
